Add files via upload
This is an updated code for the roulette wheel with a speaker and it detecting if a clap was detected. If a clap is detected, a specific sound is made and if not, a specific sound is made to signal no clap and the game is over, the final score is displayed and you must clap to restart.